my_langchain_project/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
langchain-core>=0.1.0
langchain-experimental>=0.0.1
langchain-huggingface
langchain-qdrant
pydantic>=2.0
qdrant-client==1.13.3
fastapi
uvicorn
jinja2
python-multipart
aiofiles
pandas
python-dotenv
requests
sentence-transformers